rename read-types.ts to bindings.ts and readTypes() to addBindings(eng)
