--- src/read-types/foo.d.ts
@@ -1,11 +1,0 @@
-declare var Foo: FooConstructor;
-
-interface Foo {
-    valueOf(): boolean;
-}
-
-interface FooConstructor {
-    new(value?: any): Foo;
-    <T>(value?: T): boolean;
-    readonly prototype: Foo;
-}
